+++ CHANGES.rst
@@ -4,6 +4,12 @@ Changelog
 0.1a2 (Unreleased)
 ------------------
 
+- Fixes problem with copy (see issue #2).
+  [davidemoro]
+
+- Added functional tests.
+  [davidemoro]
+
 
 
 0.1a1 (2014-01-24)